[games] Place icons on a white background.
Some game icons have transparent backgrounds which must be filled in
with white in order to comply with the ChromeOS app icon spec. This CL
creates the white circle background and updates the unit test
to check that an icon is set.

This does not affect icon sizing or masking.

Bug: 1350304
Change-Id: Icc8fe3c7af1905639089c18832270ad885393442
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3815112
Reviewed-by: Melissa Zhang <melzhang@chromium.org>
Commit-Queue: Rachel Wong <wrong@chromium.org>
Reviewed-by: Tony Yeoman <tby@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034304}