Cleanup animation of the GridCell activity indicator
The visibility of a GridCell's activity indicator is controlled by
setting its hidden property so there is no need to stop animation at the
same time. Not only is this unnecessary, but it could lead to race
conditions within the activity indicator implementation as it already
cleans up animation when hidden.

Additionally, in `showActivityIndicator`, move the start of animation to
after it is visible.

Bug: 1350734
Change-Id: I4f250b9d477179f5adf9042d75ec23a1927a9a39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3827981
Reviewed-by: edchin <edchin@google.com>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034345}