[Office upload] Send back user action on dialog close
Populate the dialog with the selected file.

When closing the dialog, respond with the user-selected action
(instead of using `chrome.send('dialogClose')`).

Readd cloud_upload_resources.pak to main pak.file (removed in
CL:3613280).

callback is called with the associated action and the dialog is closed.

Test: When clicking 'Cancel' and 'Upload', the CloudUploadDialog's
Bug: 1336924
Change-Id: I0d55563976c1f32b0ca85a4b04c740f902ad3872
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3817427
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034390}