[segmentation] Model factory gets default model from config
The default models will be set from the Config and the factory uses the
models from the config. Ideally the model factory is not needed since
the service code can read from config instead. But, there are lot of
tests using the factory to test the platform instead of config. So,
keeps the factory around.

BUG=1346389

Change-Id: I5872d619c911b37e79d03046d139bebc17d98c81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3795740
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Siddhartha S <ssid@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034418}