Updater: Change test configurations a bit, re-enable some tests.
IntegrationTest.SelfUpdateFromOldReal has not flaked in the past month,
according to Flake Portal, so re-enabling. Sheriffs: I'm not confident
that flake portal is correctly picking up flakes. If you see the test
failing on CI bots, and this CL is less than a week old, please revert
and link us to the failure.

This CL also disables most of the integration tests on component builds.
(Previously, they would run for Windows component builds.) Although they
technically pass on Windows, they rarely run well or under default flags
because the component builds are significantly larger and take much more
time to uncompress, install, etc. The overall position of the Chromium
project is that component builds are not portable outside of the build
directory, and so even if they work by accident, let's not commit to it.

The CL also re-enables the stress test. This test can likely be
re-enabled without causing problems after
https://chromium-review.googlesource.com/c/chromium/src/+/3712502.

Thank you!

Bug: 1341471
Change-Id: I2d561036cbc3f736af77a739d244dc38f3bb1741
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3828081
Reviewed-by: Sorin Jianu <sorin@chromium.org>
Commit-Queue: Joshua Pawlicki <waffles@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034458}