[UserPolicy]🏢 Replace the User Info API host when needed
iOS switched to new OAuth2 API that doesn't host the User Info API
anymore. This makes the User Info URL invalid with the new host.

We are currently exploring the best course of action to deal with that:
should we transition to the new API for all platforms? or should we
stick with the legacy API?

In the mean time, to be able to test User Policy on iOS, we need
something ad hoc to fix the problem.

Bug: 1352139
Change-Id: I78105c3db04def9f26d47f042fc4672927ad8c91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3826825
Reviewed-by: Owen Min <zmin@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/main@{#1034544}