Accessibility: Add an API to get DLC file contents
This change adds an extension API called
chrome.accessibilityPrivate.getDlcContents, which reads a DLC file and
returns the contents to the calling extension.

The motivation behind this API is described in both the TTS Using
Language Packs [1] and Pumpkin for Dictation [2] design documents.
More broadly, there are now multiple Accessibility extensions that
are leveraging DLCs, and as a result, need an easy way to access them.

For an example of how this API can be used, please see this
work-in-progress CL [3], which calls the API from the Google TTS
extension.

[1] https://docs.google.com/document/d/1M_uGwwWfq-g5FI1d0JPT3WXo9M2Z5c0hK-5cPNwEOTY/edit?usp=sharing&resourcekey=0-cGb_-SIBLBQyQpMLVPqktw
[2] https://docs.google.com/document/d/1nZxo9rG0kXKbI4ee_dSudxFWhsciazHIN-CZGgkjBsE/edit?usp=sharing
[3] https://critique.corp.google.com/cl/463904872

AX-Relnotes: N/A
Bug: b:213337877
Change-Id: Iab965f8529380d61f46a64b723fb9011406a9525
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3787009
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034549}