Iframe feed content in the sidepanel.
Iframe-ing the feed content avoids fetching a bootstrap script from GWS that then fetches the feed content and doc.writes() it into the frames DOM.

This change also modifies the feature flag to use a url for the feed content instead of the script, and makes the prod endpoint the default. This flag is also added to about_flags.

Bug: 1292623, b/235878510
Change-Id: Iaaf01ef102dad9c1b30df87f260c68d5e812c3d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3825987
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Commit-Queue: Jonathan Freed <freedjm@google.com>
Cr-Commit-Position: refs/heads/main@{#1034602}