[ViewsAX] Encapsulate logic of AXTreeManagerMap behind AXTreeManager
In this CL, we encapsulate the logic related to the
AXTreeManagerMap in the AXTreeManager. The latter class is now
solely responsible for registering a manager in the map, updating it
and removing it. It now owns a static instance of the map, and the
map is not a singleton anymore.

From now on, we can only get an AXTreeManager using one of the two following static functions:
* AXTreeManager::FromID
* AXTreeManager::ForChildTree

We don't expect any extending classes to access the AXTreeManagerMap instance directly, so we made that attribute private on the root
class. The only exception is the TestAXTreeManager, but since we only
use that class in tests, we made this class a friend of the root
class.

Bug: 1049261
Change-Id: I6e788138c3f22b08f232555248b17aa8893e6f09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3812485
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1034610}