blink: makes use of enum for html tag names
This is a patch in a series of patches to make the parsing code
use enums for html tags rather than strings.

This patch adds HTMLTokenName, which contains the local name of
a QualifiedName, and if the name is a valid html tag, an HTMLTag.
AtomicHTMLToken has an HTMLTokenName, as does HTMLStackItem.

This also converts some places to use the HTMLTag where it makes sense.

BUG=1351151
TEST=covered by tests

Change-Id: I18d2c696666de6e75e7d70869f9c8bcc776673a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3816989
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034615}