[WebContentsImpl] Don't call DidStartNavigation in child frames
Navigations in child frames are major contributors to scroll jank.
In main and child frames, DidStartNavigation is called, which calls ~20
java observers. However, all of these observers are no-ops if in a child
frame navigation, so we don't have to notify observers at all if we're
in a child frame (saves ~1ms per navigation on pixel5 debug build)

List of all observers and what they do:
https://docs.google.com/spreadsheets/d/1MySl5qeeQWpH0N3YPm_53LP2SbkY2_gLOvVGyigG0sc/edit?usp=sharing

Bug: 1337446
Change-Id: Id59b5d2f02743bfbc594cc7dd716b7df60cb40df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3758675
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Peilin Wang <peilinwang@google.com>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034621}