Stop ref'ing canvas resources in FrameResource
This change stops tracking CanvasResource objects in FrameResource
until the resource is returned by the placeholder canvas. This ensures
that the reference to CanvasResource is unique when it is transferred.  This is a pre-requisite for respecting checks that will prevent
concurrent resource access in a follow-up CL.

Bug: 1348616
Change-Id: I53cce0e05167ce18b424391c6743aca34f294c93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3826928
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034648}