[ViewsAX] Remove unnecessary AXTreeManager::ToString
In this CL, we remove the AXTreeManager::ToString function. This
function was only used once in AXPosition to log the AXTreeData (via
AXTreeData::ToString) and, now that we added a public reference to
the AXTree on the AXTreeManager class, we are able to inline that
call.

Even though it was called AXTreeManager::ToString, it was actually
ToString-ing the AXTreeData, not the AXTreeManager. If we ever have a
need to log more things than just the tree data, we should add back
this function with a richer implementation.

Bug: 1049261
Change-Id: Iaadaca06b13bdcf955ff0ea10adeddf062613a56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3821443
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1034698}