[basic-card] Remove basic card code from PaymentRequestUI
Basic card is now fully deprecated and disabled, and we are removing
its code from the codebase. This CL deletes the basic-card code from
PaymentRequestUI. As a consequence, it also removes code from
PaymentRequestTestRule that was still being 'used' (compiled) in a
number of disabled tests. Most are removed in this CL as they were
basic-card specific, but one is left with a TODO to port the test away
from relying on basic-card in the future.

Bug: 1209835
Change-Id: I537ea22095ded9a12f66106b0b0a47959bbd6564
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3827884
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034701}