[ZPS] Add feature flag to toggle Zero-Prefix Suggest prefetching on Web
In order to pave the way for enabling ZPS prefetching on the Web
(non-NTP and non-SRP URLs), this CL adds a new feature flag which
is intended to enable/disable ZPS prefetch logic on the Web. Please
refer to the linked bug for more context regarding this change.

Bug: 1344004
Change-Id: I6a05c2fd668a8a605f3d9280cadab60bbd65a84b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3827170
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Khalid Peer <khalidpeer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034741}