[Files app] Refactor DiskMountManager::MountPointMap -> MountPoints
Changed the type of the DiskMountManager::MountPointMap collection from
a map<string, MountInfo> to a set<MountInfo> indexed by mount path.

Previously, the mount path string was duplicated and stored twice (once
as key of the map entry, and once as part of the associated MountInfo
value). Now, with the set, there is no such data duplication, and the
code using this collections looks simpler and more natural.

BUG=chromium:1313346
TEST=autoninja -C out/Default unit_tests browser_tests components_unittests ash_components_unittests chrome

Change-Id: Ic5dc1b485af005f6648e11e8600c3b6613f99a8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3828330
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Marcello Salomao <msalomao@google.com>
Cr-Commit-Position: refs/heads/main@{#1034888}