dpwas: Fix PWA hang for few seconds when opening new window
For bug 1351631, we found that this is because the remote mojo object of mojom::RenderWidgetHostNSViewHost and mojom::WebContentsNSViewHost haven't been bind to the ResizeHelper task runner. So that the mojo disconnected handler cannot interrupt the ResizeHelper blocking.

When new window created, ResizeHelper will block the thread to wait for the first frame arrive. But because of associated mojo connection, the SetCALayer mojo call (new frame is from this mojo call) sometimes not allowed to be executed before the mojo disconnected handler. And disconnected handler cannot interupt the ResizeHelper blocking. So it gets into a deadlock. We must wait for the ResizeHelper stopping blocking on it's own after 5 seconds.

To fix this, we bind mojom::RenderWidgetHostNSViewHost and mojom::WebContentsNSViewHost on the ResizeHelper task runner.

Manually tested:
(1) Open this https://pwab-ug-repro.vercel.app and install the PWA app.
(2) Open the PWA app.
(3) Click the "Click to open new window and repro the bug" multiple times.
(4) No hang should be found.

Bug: 1351631
Change-Id: I11dbe5d55b154a84f3185f77ddf41283c675c52c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3829004
Commit-Queue: Samuel Qian <Qian.Zheqi@microsoft.com>
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034946}