terminal: Don't default the container name to "penguin"
Currently if the container name is empty or unset the terminal tries
to connect to container "penguin", which is the name of the default
crostini container. However, this means it is impossible to have the
terminal use vsh to connect to a VM instead of a container.

This is required for the bruschetta install flow, so remove this
behavior. Instead, treat an empty container name as no container.

Bug: b:233289313
Change-Id: I7cde410b39c638237f697ad1f69368ae8d83a81d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3826335
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Auto-Submit: Fergus Dall <sidereal@google.com>
Commit-Queue: Fergus Dall <sidereal@google.com>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034987}