Add UMA stats for the content of Android /data directory
To prepare for the virtio-blk /data migration
(go/arc:virtio-blk-migration) and the "Play files" fixup
(go/arc-r-play-files-permission-fixup), this CL adds UMA stats about the
number of files and the size of directories in some subdirectories of
Android /data. See http://b/240229355#comment3 for the full list of what
to be measured.

Bug: b:240229355, b:239934067
Test: Check the results in chrome://histograms.

Change-Id: I15011176bf99ca2f6788ab09e09fd60d2f880728
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3802296
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Risan <risan@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1034988}