PageLoadMetrics: Prerender2 support for DocumentWritePageLoadMetricsObserver
This class measures effect of `document.write()` on parsing and FCP.
As `document.write()` is strongly discouraged [1], we think it is
enough to record non prerendered case and this class doesn't support
prerendering.

This CL makes it clear to use STOP_OBSERVING and adds comment.

[1] https://html.spec.whatwg.org/multipage/dynamic-markup-insertion.html#dom-document-write-dev

Bug: 1317494
Change-Id: I43896770bafdfa0656be20306b1f7dc2ea87d24a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3817674
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Ken Okada <kenoss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1035022}