Remove SetPlatformTrack calls from MediaStreamComponent::clone()
Instead of creating a MediaStreamComponent then immediately setting a
PlatformTrack on it via SetPlatformTrack (or inside
MediaStreamAudioSource::ConnectToTrack), create it first and provide it
in the MSC constructor.

Part of a broad refactoring effort to align the lifecycles of MSC and
the PlatformTrack instances, eventually combining the two classes into
one.

Bug: 1302689
Change-Id: I8514810e16a620e0f3c91e76e4c9714d44456fa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3827423
Commit-Queue: Tony Herre <toprice@chromium.org>
Reviewed-by: Tove Petersson <tovep@chromium.org>
Auto-Submit: Tony Herre <toprice@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1035033}