PSCS: Queue updates sent while RFH is pending commit
It is possible for
ContentSettingsManager.(ContentBlocked|StorageAccessed) to be called
in between ReadyToCommitNavigation and DidFinishNavigation, at which
point we won't have a PageSpecificContentSettings object created for
the page. This is currently causing test flakiness (see bug). To fix
this, we queue any updates sent at this point, and resend them after
we create a PSCS.

Bug: 1348048
Change-Id: I2770da239ed8f3631636cb19db6d5a651c8aa04a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3817781
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Commit-Queue: Adithya Srinivasan <adithyas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1035042}