[WebContentsObserverProxy] Add metric for didStartNavigation
After renaming didStartNavigation to didStartNavigationInPrimaryMainFrame
and calling it only on the primary main frame, we have an experiment to
notify java observers spuriously on child frames. With this metric, we
want to see how much time we save by only notifying java on the primary
main frame.

Bug: 1337446
Change-Id: Iab2e49a5ed51c2341d805131918c06965827a472
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3827078
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Peilin Wang <peilinwang@google.com>
Cr-Commit-Position: refs/heads/main@{#1035061}