Default-enable global desktop site setting for premium tablets - part 2
Implement a message to notify the user and provide an option to opt-out of the default desktop site global setting update.
Screenshot: https://screenshot.googleplex.com/87MQYyWFEovVVMq.png

On FRE launch of a ChromeTabbedActivity (CTA), the message will be prioritized after the existing first-run prompts (last in this list being the LanguageAsk prompt) and before the blue tooltips. On CCT, the message will be triggered in a deferred startup task.

This can cause a delay until subsequent launches in showing the message in CTA if higher-priority eligible prompts are triggered, and to track if the message needs to be shown, a SharedPreference key will be added when the setting is default-enabled and subsequently removed once the message is shown, since the message is required to be shown at most once.

Bug: 1343916
Change-Id: Ib9ad043ca191361f477cf538d6ae7130136a44a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3812362
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Aishwarya Rajesh <aishwaryarj@google.com>
Reviewed-by: Shu Yang <shuyng@google.com>
Cr-Commit-Position: refs/heads/main@{#1035063}